+++ app/src/main/java/open/furaffinity/client/adapter/profileSectionsPagerAdapter.java
@@ -11,10 +11,10 @@ import androidx.fragment.app.FragmentPagerAdapter;
 import open.furaffinity.client.R;
 import open.furaffinity.client.fragmentTabsNew.manageAccountSettings;
 import open.furaffinity.client.fragmentTabsNew.manageBadges;
-import open.furaffinity.client.fragmentTabsOld.manageUserSettings;
+import open.furaffinity.client.fragmentTabsNew.manageUserSettings;
 import open.furaffinity.client.fragmentTabsNew.notImplementedYet;
 import open.furaffinity.client.fragmentTabsNew.manageAvatar;
-import open.furaffinity.client.fragmentTabsOld.manageWatches;
+import open.furaffinity.client.fragmentTabsNew.manageWatches;
 import open.furaffinity.client.fragmentTabsNew.manageContactInfo;
 import open.furaffinity.client.fragmentTabsNew.manageFolders;
 import open.furaffinity.client.fragmentTabsNew.manageJournals;
