+++ app/src/main/java/open/furaffinity/client/adapter/stringListAdapter.java
@@ -16,7 +16,7 @@ import java.util.List;
 import open.furaffinity.client.R;
 import open.furaffinity.client.activity.mainActivity;
 import open.furaffinity.client.fragmentDrawers.journal;
-import open.furaffinity.client.fragmentTabsOld.msgPmsMessage;
+import open.furaffinity.client.fragmentTabsNew.msgPmsMessage;
 import open.furaffinity.client.fragmentDrawers.user;
 import open.furaffinity.client.fragmentDrawers.view;
