+++ app/src/main/java/open/furaffinity/client/adapter/userSectionsPagerAdapter.java
@@ -14,8 +14,8 @@ import open.furaffinity.client.fragmentTabsNew.notImplementedYet;
 import open.furaffinity.client.fragmentTabsOld.shouts;
 import open.furaffinity.client.fragmentTabsOld.userGallery;
 import open.furaffinity.client.fragmentTabsOld.userJournals;
-import open.furaffinity.client.fragmentTabsOld.userProfile;
-import open.furaffinity.client.fragmentTabsOld.watch;
+import open.furaffinity.client.fragmentTabsNew.userProfile;
+import open.furaffinity.client.fragmentTabsNew.watch;
 import open.furaffinity.client.fragmentTabsNew.webViewContent;
 import open.furaffinity.client.pages.user;
 import open.furaffinity.client.utilities.messageIds;
