+++ app/src/main/java/open/furaffinity/client/fragmentTabsNew/manageUserSettings.java
@@ -1,19 +1,9 @@
-package open.furaffinity.client.fragmentTabsOld;
+package open.furaffinity.client.fragmentTabsNew;
 
-import android.os.AsyncTask;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.RadioButton;
 import android.widget.Spinner;
 import android.widget.Toast;
-
-import androidx.fragment.app.Fragment;
-
-import java.util.HashMap;
-import java.util.concurrent.ExecutionException;
 
 import open.furaffinity.client.R;
 import open.furaffinity.client.abstractClasses.abstractPage;
@@ -21,11 +11,8 @@ import open.furaffinity.client.pages.controlsUserSettings;
 import open.furaffinity.client.utilities.fabCircular;
 import open.furaffinity.client.utilities.kvPair;
 import open.furaffinity.client.utilities.uiControls;
-import open.furaffinity.client.utilities.webClient;
 
-public class manageUserSettings extends Fragment {
-    private static String TAG = manageUserSettings.class.getName();
-
+public class manageUserSettings extends open.furaffinity.client.abstractClasses.tabFragment {
     private RadioButton accept_trades_yes;
     private RadioButton accept_trades_no;
     private RadioButton accept_commissions_yes;
@@ -34,12 +21,16 @@ public class manageUserSettings extends Fragment {
 
     private fabCircular fab;
 
-    private open.furaffinity.client.utilities.webClient webClient;
     private controlsUserSettings page;
 
     private boolean isLoading = false;
 
-    private void getElements(View rootView) {
+    @Override
+    protected int getLayout() {
+        return R.layout.fragment_manage_user_settings;
+    }
+
+    protected void getElements(View rootView) {
         accept_trades_yes = rootView.findViewById(R.id.accept_trades_yes);
         accept_trades_no = rootView.findViewById(R.id.accept_trades_no);
         accept_commissions_yes = rootView.findViewById(R.id.accept_commissions_yes);
@@ -51,7 +42,7 @@ public class manageUserSettings extends Fragment {
         fab.setVisibility(View.GONE);
     }
 
-    private void fetchPageData() {
+    protected void fetchPageData() {
         if(!isLoading) {
             isLoading = true;
             page = new controlsUserSettings(page);
@@ -59,9 +50,7 @@ public class manageUserSettings extends Fragment {
         }
     }
 
-    private void initPages() {
-        webClient = new webClient(requireContext());
-
+    protected void initPages() {
         page = new controlsUserSettings(getActivity(), new abstractPage.pageListener() {
             @Override
             public void requestSucceeded(abstractPage abstractPage) {
@@ -92,44 +81,17 @@ public class manageUserSettings extends Fragment {
         });
     }
 
-    private void updateUIElementListeners(View rootView) {
-        fab.setOnClickListener(new View.OnClickListener() {
+    protected void updateUIElementListeners(View rootView) {
+        fab.setOnClickListener(v -> new open.furaffinity.client.submitPages.submitControlsUserSettings(getActivity(), new abstractPage.pageListener() {
             @Override
-            public void onClick(View v) {
-                HashMap<String, String> params = new HashMap<>();
-                params.put("do", "update");
-                params.put("key", page.getKey());
-                params.put("accept_trades", ((accept_trades_yes.isChecked()) ? ("1") : ("0")));
-                params.put("accept_commissions", ((accept_commissions_yes.isChecked()) ? ("1") : ("0")));
-                params.put("featured_journal_id", ((kvPair) featured_journal_id.getSelectedItem()).getKey());
-                params.put("save_settings", "Save Settings");
-
-                try {
-                    new AsyncTask<webClient, Void, Void>() {
-                        @Override
-                        protected Void doInBackground(open.furaffinity.client.utilities.webClient... webClients) {
-                            webClients[0].sendPostRequest(open.furaffinity.client.utilities.webClient.getBaseUrl() + controlsUserSettings.getPagePath(), params);
-                            return null;
-                        }
-                    }.execute(webClient).get();
-                } catch (ExecutionException | InterruptedException e) {
-                    Log.e(TAG, "Could not update user settings: ", e);
-                }
+            public void requestSucceeded(abstractPage abstractPage) {
+                Toast.makeText(getActivity(), "Successfully updated user settings", Toast.LENGTH_SHORT).show();
             }
-        });
-    }
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_manage_user_settings, container, false);
-        getElements(rootView);
-        initPages();
-        fetchPageData();
-        updateUIElementListeners(rootView);
-        return rootView;
+            @Override
+            public void requestFailed(abstractPage abstractPage) {
+                Toast.makeText(getActivity(), "Failed to update user settings", Toast.LENGTH_SHORT).show();
+            }
+        }, page.getKey(), ((accept_trades_yes.isChecked()) ? ("1") : ("0")), ((accept_commissions_yes.isChecked()) ? ("1") : ("0")), ((kvPair) featured_journal_id.getSelectedItem()).getKey()).execute());
     }
 }
