+++ app/src/main/res/layout/fragment_msgpmsmessage.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragmentTabsOld.msgPmsMessage">
+    tools:context=".fragmentTabsNew.msgPmsMessage">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
